versionsDefinitives

+++ Sprint3Tasca2/lib/src/main/java/n1ex1/Observable.java
@@ -5,6 +5,6 @@ public interface Observable {
 
 	public void detach(Observer o);
 
-	public void notifyChange(float novaPuntuacio);
+	public void notifyChange(double novaPuntuacio);
 
 }
